requirements.txt
a2a-sdk[http-server]==0.3.22
httpx==0.28.1
googlemaps==4.10.0
Pillow==12.1.0
requests==2.32.5
polyline==2.0.4
uvicorn==0.40.0
